Make conversion explicit (clang trunk warning).

+++ sql/odbc/src/TODBCStatement.cxx
@@ -637,8 +637,8 @@ const char* TODBCStatement::ConvertToString(Int_t npar)
       case SQL_C_SLONG:   snprintf(buf, 100, "%d", *((SQLINTEGER*) addr)); break;
       case SQL_C_ULONG:   snprintf(buf, 100, "%u", *((SQLUINTEGER*) addr)); break;
 #else
-      case SQL_C_SLONG:   snprintf(buf, 100, "%ld", *((SQLINTEGER*) addr)); break;
-      case SQL_C_ULONG:   snprintf(buf, 100, "%lu", *((SQLUINTEGER*) addr)); break;
+      case SQL_C_SLONG:   snprintf(buf, 100, "%ld", (long)*((SQLINTEGER*) addr)); break;
+      case SQL_C_ULONG:   snprintf(buf, 100, "%lu", (unsigned long)*((SQLUINTEGER*) addr)); break;
 #endif
       case SQL_C_SBIGINT: snprintf(buf, 100, "%lld", *((Long64_t*) addr)); break;
       case SQL_C_UBIGINT: snprintf(buf, 100, "%llu", *((ULong64_t*) addr)); break;
